creating-features-in-postgis-using-openlayers-draw-interactions-in-alpinejs-and-livewire

+++ app/Console/Commands/ReloadMonuments.php
@@ -2,6 +2,7 @@
 
 namespace App\Console\Commands;
 
+use App\Models\User;
 use App\Models\Monument;
 use Illuminate\Console\Command;
 use Illuminate\Support\Facades\DB;
@@ -29,15 +30,24 @@ class ReloadMonuments extends Command
      */
     public function handle()
     {
+        $user = User::first();
+
+        if (!$user) {
+            $this->info('No user found in the database, the monuments cannot be loaded.');
+
+            return 0;
+        }
+
         $geojson = file_get_contents(resource_path('/geodata/geojson/monuments.geojson'));
 
         Monument::truncate();
 
-        $features = collect(json_decode($geojson, true)['features'])->each(function ($feature) {
+        $features = collect(json_decode($geojson, true)['features'])->each(function ($feature) use ($user) {
             Monument::create([
                 'name' => $feature['properties']['name'],
-                'image' => $feature['properties']['image'],
+                'image' => null,
                 'geom' => DB::raw("ST_GeomFromGeoJSON('" . json_encode($feature['geometry']) . "')"),
+                'user_id' => $user->id
             ]);
         });
